ris_backend/requirements.txt
absl-py==2.0.0
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
attrs==23.2.0
bcrypt==4.0.1
bitarray==2.9.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
coverage==7.5.4
customtkinter==5.2.2
cycler==0.12.1
cytoolz==0.12.3
darkdetect==0.8.0
distlib==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
fastapi==0.103.1
filelock==3.13.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.0
gast==0.5.4
gitdb==4.0.11
GitPython==3.1.43
glob2==0.7
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
hexbytes==0.3.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hugchat==0.4.8
huggingface-hub==0.23.4
idna==3.6
imageio==2.33.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lru-dict==1.2.0
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.10.9
ml-dtypes==0.2.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
npzviewer==0.2.0
numpy==1.26.2
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.10.0.84
opt-einsum==3.3.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parsimonious==0.9.0
passlib==1.7.4
Pillow==10.1.0
platformdirs==4.2.2
pluggy==1.5.0
protobuf==4.25.3
psutil==6.0.0
py-cpuinfo==9.0.0
py-solc-x==2.0.2
pyarrow==16.1.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybind11==2.12.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.1.1
pyproject-api==1.7.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.12
PyQt5-sip==12.13.0
pytest==8.2.2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
pyunormalize==15.1.0
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rlp==4.0.0
rpds-py==0.18.0
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.6
SQLAlchemy==2.0.28
starlette==0.27.0
streamlit==1.36.0
sympy==1.12
tenacity==8.4.1
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.34.0
tensorflow-macos==2.15.0
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2023.12.9
toml==0.10.2
toolz==0.12.1
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tornado==6.4.1
tox==4.15.1
tqdm==4.66.1
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.10.0
-e git+https://github.com/hiimjupter/YOLOv10_Finetuning_ProtectionHelmetDetection@ac74cc0c4660f571676bf7b4f559072b99cedc72#egg=ultralytics
urllib3==2.1.0
uuid==1.30
uvicorn==0.27.1
uvloop==0.19.0
virtualenv==20.26.3
watchfiles==0.22.0
web3==6.15.1
websockets==12.0
Werkzeug==3.0.1
wget==3.2
wrapt==1.14.1
yarl==1.9.4
